update to work on osx

+++ install.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 set -e
 
-sudo apt-get install -y git vim zsh
+if [ -f /etc/lsb-release ]; then
+	sudo apt-get install -y git vim zsh
+fi
 git clone --recursive git@github.com:robstrong/dotfiles ~/.dotfiles
 cd ~/.dotfiles
 git init
@@ -11,8 +13,10 @@ vim +PluginInstall +qall
 ln -s ~/.dotfiles/vim/vimrc ~/.vimrc
 ln -s ~/.dotfiles/zsh/zshrc ~/.zshrc
 ln -s ~/.dotfiles/git/gitconfig ~/.gitconfig
-# allow neovim config as well
-ln -s ~/.vim ~/.config/nvim
-ln -s ~/.vimrc ~/.config/nvim/init.vim
+if [ -d ~/.config ]; then
+    # allow neovim config as well
+    ln -s ~/.vim ~/.config/nvim
+    ln -s ~/.vimrc ~/.config/nvim/init.vim
+fi
 
 chsh -s $(which zsh)
